requirements.txt
Flask==3.0.3
flask_marshmallow==1.2.1
flask_sqlalchemy==3.1.1
marshmallow==3.21.2
openai==1.30.3
pytest==8.2.1
python-dotenv==1.0.1
